pyproject.toml
[tool.poetry]
name = "simple-calculator-exl"                       # This is the PyPI name (can have a dash)
version = "0.1.3"
description = "A tiny example package providing simple math helpers."
license = "MIT"
authors = ["Dharmender Singh <dharmender071999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.50.0"
streamlit-aggrid = "^1.1.9"
fastmcp = "^2.12.4"
fastapi = "^0.118.1"
google-generativeai = "^0.8.5"
snowflake-connector-python = "^3.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
ruff = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# 👇 IMPORTANT: must be an array of tables (each [[...]])
[[tool.poetry.packages]]
include = "simple_calculator_exl"
